+++ examples/resources/iosxe_system/resource.tf
@@ -1,4 +1,5 @@
 resource "iosxe_system" "example" {
   hostname             = "ROUTER-1"
+  ip_routing           = true
   ipv6_unicast_routing = true
 }