+++ internal/provider/resource_iosxe_system_test.go
@@ -17,6 +17,7 @@ func TestAccIosxeSystem(t *testing.T) {
 				Config: testAccIosxeSystemConfig_all(),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr("iosxe_system.test", "hostname", "ROUTER-1"),
+					resource.TestCheckResourceAttr("iosxe_system.test", "ip_routing", "true"),
 					resource.TestCheckResourceAttr("iosxe_system.test", "ipv6_unicast_routing", "true"),
 				),
 			},
@@ -40,6 +41,7 @@ func testAccIosxeSystemConfig_all() string {
 	return `
 	resource "iosxe_system" "test" {
 		hostname = "ROUTER-1"
+		ip_routing = true
 		ipv6_unicast_routing = true
 	}
 	`