+++ docs/guides/changelog.md
@@ -16,6 +16,7 @@ description: |-
 - Add iosxe_dhcp resource and data source
 - Add iosxe_msdp resource and data source
 - Add iosxe_msdp_vrf resource and data source
+- Add ip_routing attribute to iosxe_system resource
 
 ## 0.1.5